requirements.txt
openai
pinecone-client
PyPDF2
pandas
sentence-transformers
streamlit
gradio